Rename and refactor tests TestSites -> TestJumps/TestTransitions
